[fix] Improve builder to include adding list of metrics
PROBLEM:
Currently, the user has to use WithCollector() function to
register each metric. This makes registering of large number
of metrics difficult.

SOLUTION:
This PR provides an additional WithCollectors() function which
takes in a list of metrics which are to be registered with
prometheus. The user, can hence directly provide a list of metrics
to register.